clear all tests for suboctave issue with mod duo, but did not find this fucking fucking bug

+++ src/Ramp.cpp
@@ -615,7 +615,7 @@ float Ramp::get_octave_image_value(float speed, bool leaving){
         int delta = period_count % int(1/abs(speed)); /* speed can't be 0 here */
         if (delta != 0){
             if (speed > 0.0f){
-                value += (audio_memory[uint32_t(frame+1)] - audio_memory[uint32_t(frame)]) * delta * speed;
+                value += (audio_memory[frame+1] - audio_memory[frame]) * delta * speed;
             } else {
                 int compare_frame = frame -1;
                 if (compare_frame < 0){
